// src/models/Category.ts
enum Category {
  Comedy = "Comedy",
  Thriller = "Thriller",
  Horror = "Horror",
  Action = "Action",
  Anime = "Anime",
  Drama = "Drama",
  SciFi = "Sci-Fi",
  Fantasy = "Fantasy",
  Documentary = "Documentary",
  Romance = "Romance",
  Crime = "Crime",
  Adventure = "Adventure",
  Family = "Family",
  Musical = "Musical",
  Biography = "Biography",
  Historical = "Historical",
  Sports = "Sports",
  Animation = "Animation",
  Kids = "Kids",
  Educational = "Educational",
  StandUp = "Stand-Up"
}

export default Category;